add startup func for kafka and test
