+++ src/kafka/mod.rs
@@ -1,1 +1,2 @@
 pub mod consumer;
+pub mod startup;
